+++ test/fixtures/main.ts
@@ -1,7 +1,9 @@
-import { SerialPort } from 'serialport'
+// TODO: https://github.com/serialport/node-serialport/issues/2464
+// import * as serialport from 'serialport'
 import sqlite3 from 'sqlite3'
+import * as fsevents from 'fsevents'
 
 export {
-  SerialPort,
   sqlite3,
+  fsevents,
 }