+++ vite.config.ts
@@ -5,12 +5,9 @@ import { builtinModules } from 'node:module'
 import { defineConfig } from 'vite'
 import pkg from './package.json'
 
-const isdev = process.argv.slice(2).includes('--watch')
-
 export default defineConfig({
   build: {
     minify: false,
-    emptyOutDir: !isdev,
     lib: {
       entry: 'src/index.ts',
       formats: ['cjs', 'es'],
@@ -18,7 +15,6 @@ export default defineConfig({
     },
     rollupOptions: {
       external: [
-        'electron',
         'vite',
         ...builtinModules,
         ...builtinModules.map(m => `node:${m}`),